% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{academic_year}
\alias{academic_year}
\title{Format a Date/date-like column into an academic-year label, e.g. "2026-2027"}
\usage{
academic_year(d)
}
\arguments{
\item{d}{Date vector}
}
\description{
Format a Date/date-like column into an academic-year label, e.g. "2026-2027"
}
