% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{AMION_START_AY}
\alias{AMION_START_AY}
\title{Earliest academic year Amion data is fetched from by default. Bump this
if older history is needed.}
\usage{
AMION_START_AY
}
\description{
Earliest academic year Amion data is fetched from by default. Bump this
if older history is needed.
}
