% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_category.R
\name{classify_super_category}
\alias{classify_super_category}
\title{Roll up a vector of categories (from classify_rotation() or
classify_session()) into the 7 super-categories. Anything not in
SUPER_CATEGORY_MAP comes back as "UNMAPPED".}
\usage{
classify_super_category(category)
}
\arguments{
\item{category}{Character vector of category values.}
}
\description{
Roll up a vector of categories (from classify_rotation() or
classify_session()) into the 7 super-categories. Anything not in
SUPER_CATEGORY_MAP comes back as "UNMAPPED".
}
