% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duty_hour_summary.R
\name{.dt_from_hhmm}
\alias{.dt_from_hhmm}
\title{HHMM (character/integer) + Date -> POSIXct, vectorized, NA-safe.}
\usage{
.dt_from_hhmm(date, hhmm)
}
\description{
HHMM (character/integer) + Date -> POSIXct, vectorized, NA-safe.
}
\keyword{internal}
