% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_attendance.R
\name{EXPECTED_CONFERENCE_MAP}
\alias{EXPECTED_CONFERENCE_MAP}
\title{Amion rotation category (from \code{classify_rotation()}) -> expected
conference site. "SLUH" / "VA" / "None".}
\usage{
EXPECTED_CONFERENCE_MAP
}
\description{
Amion rotation category (from \code{classify_rotation()}) -> expected
conference site. "SLUH" / "VA" / "None".
}
