% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_duty_hour_page.R
\name{mod_duty_hour_page_server}
\alias{mod_duty_hour_page_server}
\title{Duty Hours page server}
\usage{
mod_duty_hour_page_server(id, resident_id, rdm_token, redcap_url)
}
\arguments{
\item{id}{Shiny module id — must match \code{mod_duty_hour_page_ui}'s.}

\item{resident_id}{Reactive returning the current resident's record_id.}

\item{rdm_token}{RDM REDCap API token.}

\item{redcap_url}{REDCap base URL.}
}
\description{
Duty Hours page server
}
