% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{ROTATION_CATEGORY_MAP}
\alias{ROTATION_CATEGORY_MAP}
\title{Amion \code{Grouping} -> top-level rotation category.}
\usage{
ROTATION_CATEGORY_MAP
}
\description{
Amion \code{Grouping} -> top-level rotation category.
}
