% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_category.R
\name{SUPER_CATEGORY_MAP}
\alias{SUPER_CATEGORY_MAP}
\title{Category (from either classify_rotation() or classify_session()) ->
super-category. Confirmed with Fred 2026-08-16.}
\usage{
SUPER_CATEGORY_MAP
}
\description{
Category (from either classify_rotation() or classify_session()) ->
super-category. Confirmed with Fred 2026-08-16.
}
